feat(content-answers): fixed scrolling issue
