+++ src/features/content-answers/ContentAnswersModalContent.scss
@@ -1,6 +1,6 @@
 @import '../../styles/variables';
 
-.bdl-ContentAnswersModalChat {
+.bdl-ContentAnswersModalContent {
     flex: 1 1 auto;
     width: auto;
     padding: 0 $bdl-grid-unit * 5 $bdl-grid-unit * 5 $bdl-grid-unit * 5;
