+++ src/features/content-answers/__tests__/ContentAnswersModalContent.test.tsx
@@ -1,8 +1,14 @@
 import React from 'react';
-import { fireEvent, render, waitFor } from '@testing-library/react';
+import { render, screen, waitFor } from '@testing-library/react';
 
 import ContentAnswersModalContent from '../ContentAnswersModalContent';
-import { mockCurrentUser, mockFile, mockQuestionsWithAnswer } from '../__mocks__/mocks';
+import {
+    mockCurrentUser,
+    mockFile,
+    mockQuestionsNoAnswer,
+    mockQuestionsWithAnswer,
+    mockQuestionsWithAnswerAndNoAnswer,
+} from '../__mocks__/mocks';
 
 describe('features/content-answers/ContentAnswersModalContent', () => {
     const scrollIntoViewMock = jest.fn();
@@ -21,9 +27,42 @@ describe('features/content-answers/ContentAnswersModalContent', () => {
         HTMLElement.prototype.scrollIntoView = scrollIntoViewMock;
     });
 
-    afterEach(() => {
-        // Restore the original implementation of scrollIntoView
-        delete HTMLElement.prototype.scrollIntoView;
+    test('should render welcome message', () => {
+        renderComponent();
+
+        expect(screen.queryByTestId('content-answers-welcome-message')).toBeInTheDocument();
+    });
+
+    test('should render only the prompt while loading', () => {
+        renderComponent({ isLoading: true, questions: mockQuestionsNoAnswer });
+
+        expect(screen.getByText(mockQuestionsNoAnswer[0].prompt)).toBeInTheDocument();
+
+        const loadingElement = screen.getByTestId('LoadingElement');
+        expect(loadingElement).toBeInTheDocument();
+    });
+
+    test('should render only the prompt, answer and not loading', () => {
+        renderComponent({ questions: mockQuestionsWithAnswer });
+        const { answer = '', prompt } = mockQuestionsWithAnswer[0];
+
+        expect(screen.getByText(prompt)).toBeInTheDocument();
+        expect(screen.getByText(answer)).toBeInTheDocument();
+
+        const loadingElement = screen.queryByTestId('LoadingElement');
+        expect(loadingElement).not.toBeInTheDocument();
+    });
+
+    test('should render previous question with another new prompt without answer while loading', () => {
+        renderComponent({ isLoading: true, questions: mockQuestionsWithAnswerAndNoAnswer });
+        const { answer = '', prompt } = mockQuestionsWithAnswerAndNoAnswer[0];
+
+        expect(screen.getByText(prompt)).toBeInTheDocument();
+        expect(screen.getByText(answer)).toBeInTheDocument();
+        expect(screen.getByText(mockQuestionsWithAnswerAndNoAnswer[1].prompt)).toBeInTheDocument();
+
+        const loadingElement = screen.queryAllByTestId('LoadingElement');
+        expect(loadingElement.length).toEqual(1);
     });
 
     test('handleScrollToBottom is called with instant behavior on component mount', async () => {
@@ -48,26 +87,5 @@ describe('features/content-answers/ContentAnswersModalContent', () => {
 
         expect(scrollIntoViewMock).toBeCalledWith({ behavior: 'smooth' });
         expect(scrollIntoViewMock).toBeCalledTimes(2);
-    });
-
-    test('should not call scrollIntoView if scroll distance is higher than threshold the next time an answer is updated', () => {
-        const { container, rerender } = renderComponent({ questions: [] });
-        const modalContent = container.firstChild;
-
-        // @ts-ignore should be here
-        modalContent.scrollTop = -10;
-        // @ts-ignore should be available
-        fireEvent.scroll(modalContent);
-
-        rerender(
-            <ContentAnswersModalContent
-                currentUser={mockCurrentUser}
-                fileName={mockFile.name}
-                isLoading={false}
-                questions={mockQuestionsWithAnswer}
-            />,
-        );
-
-        expect(scrollIntoViewMock).not.toBeCalled();
     });
 });
